Move namespace and rename API request client
